requirements.txt
playwright>=1.40.0
pandas>=2.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0

